+++ R/rangl-package.r
@@ -1,18 +1,30 @@
-#' Tidy tables for topological, triangulated, and other spatial data structures.
+#' Tidy tables for topological spatial data structures.
 #'
+#' The 'rangl' package helps transcend the following problems in R and GIS generally: 
+#' 
+#' \itemize{
+#'  \item coordinates beyond X and Y, or longitude and latitude
+#'  \item attributes stored where appropriate, on vertices, primitives, branches, or objects
+#'  \item topology and geometry are not conflated
+#'  \item spatial data can be properly represented as a graph of spatial primitives
+#'  \item polygons as true surfaces, not just glorified lines with a path-filling rule
+#'  \item TBD higher dimensional primitives are possible
+#'  \item TBD n-dimensional rasters with curvilinear coordinates, and the discrete-continuous distinction 
+#' }
 #'
 #' @name rangl
 #' @docType package
 #' @section I. Creation:
 #' \tabular{ll}{
-#'  \code{\link{mesh}} \tab create a triangulated mesh as tidy set of tables  \cr
+#'  \code{\link{rangl}} \tab create a rangl table set from various input types  \cr
 #'  }
 #'
 #' @section II. Plotting:
 #' \tabular{ll}{
 #'  \code{\link{globe}} \tab convert X,Y planar or angular to 3D on the surface of a globe, based on the data in longitude-latitude form \cr
-#'  \code{\link{plot.trimesh}} \tab plot 2D topology in 3D \cr
-#'  \code{\link{plot.trimesh}} \tab plot 1D topology in 3D \cr
+#'  \code{\link{plot.trimesh}} \tab plot 2D topology in 3D geometry space \cr
+#'  \code{\link{plot.linemesh}} \tab plot 1D topology in 3D geometry space \cr
+#'  \code{\link{plot.pointmesh}} \tab plot 0D topology in 3D geometry space \cr
 #' }
 #'
 NULL