+++ tests/testthat/test-basic.R
@@ -11,11 +11,11 @@ library(rworldxtra)
 library("spbabel")
 data("holey")
 test_that("holey polygons and non-holy polygons", {
-  expect_that(b <- mesh(a), is_a("trimesh"))
+  expect_that(b <- rangl(a), is_a("trimesh"))
     sph <- sp(holey)
     sph$meta$proj <- "+proj=laea +ellps=wGS84"
     
-    expect_that(  tm <- mesh(sph), is_a("trimesh"))
+    expect_that(  tm <- rangl(sph), is_a("trimesh"))
   
   
 })
@@ -23,17 +23,17 @@ test_that("holey polygons and non-holy polygons", {
 library(maptools)
 data(wrld_simpl)
 test_that("simple plot", {
- b <- mesh(a)
+ b <- rangl(a)
     plot(b)
     
-    b <- mesh(a)
+    b <- rangl(a)
     globe(b)
     
     ## check that we work without a DataFrame
-    b <- mesh(geometry(a))
+    b <- rangl(geometry(a))
       rgl::rgl.clear()
    for (i in sample(seq(nrow(wrld_simpl)), 10)) {
-     globe(mesh(wrld_simpl[i, ]))
+     globe(rangl(wrld_simpl[i, ]))
      rgl::rgl.clear()
 
  }