+++ tests/testthat/test-points.R
@@ -9,10 +9,10 @@ data(wrld_simpl)
 pts <- as(as(wrld_simpl, "SpatialLinesDataFrame"), "SpatialPointsDataFrame")
 mpts <- as(as(wrld_simpl, "SpatialLinesDataFrame"), "SpatialMultiPointsDataFrame")
 test_that("points works", {
-  expect_that(mesh(pts), throws_error("you don't really"))
+  expect_that(rangl(pts), throws_error("you don't really"))
 })
 
 
 test_that("multi-points works", {
-  expect_that(mesh(mpts), is_a("pointmesh"))
+  expect_that(rangl(mpts), is_a("pointmesh"))
 })