+++ NEWS.md
@@ -3,7 +3,7 @@
 * removed old globe() plot behaviour, this function now just converts coordinates to geocentric XYZ
 
 * added generic "mesh()" function to convert SpatialPolygons, SpatialLines, and
-rgl 'mesh3d' objects that use triangle primitives
+rgl 'mesh3d' objects (only those that use triangle primitives)
 
 * removed old "tri_mesh" function