+++ NEWS.md
@@ -2,14 +2,16 @@
 
 * mesh method for trip
 
+* fix for spbabel now means MultiPoints are mesh()-able
+
 # rangl 0.2.0
 
 * removed old globe() plot behaviour, this function now just converts coordinates to geocentric XYZ
 
 * added generic "mesh()" function to convert SpatialPolygons, SpatialLines, and
-rgl 'mesh3d' objects that use triangle primitives
+rgl 'mesh3d' objects (only those that use triangle primitives)
 
-* removed old "tri_mesh" function
+* deprecated "tri_mesh" function, replaced by mesh()
 
 * renamed package to 'rangl'